+++ internal/cmlschema/lab_group.go
@@ -44,6 +44,24 @@ func NewLabGroup(ctx context.Context, group *models.LabGroup, diags *diag.Diagno
 	return value
 }
 
+// NewLabGroupFromAssociation converts a lab-group association into a Terraform value.
+func NewLabGroupFromAssociation(ctx context.Context, assoc *models.Association, diags *diag.Diagnostics) attr.Value {
+	newGroup := LabGroupModel{
+		ID:         types.StringValue(string(assoc.ID)),
+		Permission: types.StringValue(TFGroupPermissionFromAssociationPermissions(assoc.Permissions)),
+	}
+	var value attr.Value
+	diags.Append(
+		tfsdk.ValueFrom(
+			ctx,
+			newGroup,
+			types.ObjectType{AttrTypes: LabGroupAttrType},
+			&value,
+		)...,
+	)
+	return value
+}
+
 // LabGroup returns the schema for a lab group permission nested object.
 func LabGroup() map[string]schema.Attribute {
 	return map[string]schema.Attribute{